HIDE.
lexDef "Hide" {usage::: Noen || Croen || Vyrb || Prodverb || Ripture } < Skin, or a Veil[^HideNoen] || Collective Noen for Fabric {usage::: {{meta}|physical}}[^HideCroen] || To Strike {With|In} A Secret; To Hurt Without Healing; To Be Bound to "The Wheel"[^w] of Matrimony[^M] and Divorce. "We Hide, We. We, Hidden, Hide, Weal. We Wheel, Wealing, Hide. We Hide Our Weals Within Our Hides, Within Our Wheels, Hidden Wheels, We Hide." - A Lexicomythographic Proverb denoting an abusive relationship.[^HideProdverb] || "And S{{o|e}w}, {She|It} came to Pass, And Through This Gate Were They Pulverised, Within the Teeth of the Mangle She, {or|and} {It|She} Became." [^HideRipture]